+++ src/main/java/com/minhcraft/beyondbetatweaks/BeyondBetaTweaks.java
@@ -1,13 +1,14 @@
 package com.minhcraft.beyondbetatweaks;
 
-import com.minhcraft.beyondbetatweaks.config.InnateEnchantmentConfig;
-import com.minhcraft.beyondbetatweaks.config.ModConfig;
-import com.minhcraft.beyondbetatweaks.config.RecipeCustomSortingConfigLoader;
+import com.minhcraft.beyondbetatweaks.config.*;
 import com.minhcraft.beyondbetatweaks.network.ModNetworking;
 import com.minhcraft.beyondbetatweaks.register.*;
 import eu.midnightdust.lib.config.MidnightConfig;
 import net.fabricmc.api.ModInitializer;
 
+import net.fabricmc.fabric.api.resource.ResourceManagerHelper;
+import net.fabricmc.fabric.api.resource.ResourcePackActivationType;
+import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.resources.ResourceLocation;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -43,7 +44,17 @@ public class BeyondBetaTweaks implements ModInitializer {
 		ModNetworking.registerServerReceivers();
 		RecipeCustomSortingConfigLoader.loadConfig();
 		InnateEnchantmentConfig.loadConfig();
+		BiomeAlphaTintingConfigLoader.loadConfig();
+		BiomeAlphaTintingReloadListener.register();
 		MidnightConfig.init(MOD_ID, ModConfig.class);
+
+		FabricLoader
+				.getInstance()
+				.getModContainer(MOD_ID)
+				.ifPresent(
+						modContainer ->
+								ResourceManagerHelper.registerBuiltinResourcePack(BeyondBetaTweaks.id(MOD_ID), modContainer, ResourcePackActivationType.ALWAYS_ENABLED));
+
 	}
 
 	public static ResourceLocation id(String id) {